requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
django==5.0.4
    # via
    #   -r requirements.in
    #   django-markdownx
django-markdownx==4.0.7
    # via -r requirements.in
markdown==3.6
    # via django-markdownx
pillow==10.3.0
    # via django-markdownx
sqlparse==0.4.4
    # via django
tzdata==2024.1
    # via django
